created MessageBox function and replaced window alerts with it
