+++ src/styles/playerinput.css
@@ -46,4 +46,5 @@
 .dropdownList {
     display: block;
     margin: 1rem auto;
+    width: 100%
 }
